move getAtomRegistry
