+++ packages/react/src/atoms/index.ts
@@ -11,3 +11,4 @@ export type {
 } from './types';
 export { withPropMeta, withArgMeta, getFieldMeta } from './schema-utils';
 export { createAtom, type AtomSchemaInput } from './createAtom';
+export { serializeAtoms, getAtomRegistry } from './atom-registry-utils';
