+++ packages/react/src/component-layout/index.ts
@@ -1,7 +1,10 @@
 /**
  * Component Layout: createView and getAtomRegistry for rendering no-code layout documents.
- * @packageDocumentation
  */
 
-export { createView, type CallbackRegistry, type CallbackArgNamesMap, type CreateViewOptions } from './createView';
-export { getAtomRegistry } from './getAtomRegistry';
+export {
+  createView,
+  type CallbackRegistry,
+  type CallbackArgNamesMap,
+  type CreateViewOptions,
+} from './createView';
